requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
psycopg2-binary
bcrypt>=4.0.0
python-jose[cryptography]